+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 * Switched from `file_exists()` to `is_file()` for detecting image processor and variant commands
 * Added better debug logging
+* Updated to require craftcms/cms `^3.0.0-RC1`
 
 ## 1.3.4 - 2017.12.03
 ### Changed